requirements.txt
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0
matplotlib==3.7.1
tqdm==4.65.0
gradio==4.16.0